Organise files in src and test folders
